moved RegisterForm from BaseComponents to Components
